+++ ChangeLog.md
@@ -1,7 +1,10 @@
 # Change Log
 ## 0.0.2 - 2025-08-17
 ### Added
-- Added `channel`, `thread_ts`, `event_type`, and `reaction` fields to inputs
+- Added `channel`, `message_ts`, `event_type`, and `reaction` fields to inputs
+
+### Fixed
+- Fixed an issue where the original message for reactions wasn't being properly retrieved
 
 ## 0.0.1 - 2025-08-11
 ### Added